Fix file list not splitting correctly.
The replace function only replaces the first occurence, but with a regex we can set the global flag to replace all newlines in the string.

Also, replace newlines with a blank space so file paths will be separated. Occasionally this makes array get some empty strings that we can skip in the loop.